:art: Fix #221

+++ main.go
@@ -104,21 +104,23 @@ func handleSignal(server *http.Server) {
 
 func replaceServerConf() {
 	path := "theme/sw.min.js.tpl"
-	data, err := ioutil.ReadFile(path)
-	if nil != err {
-		logger.Fatal("read file [" + path + "] failed: " + err.Error())
-	}
-	content := string(data)
-	content = strings.Replace(content, "http://server.tpl.json", model.Conf.Server, -1)
-	content = strings.Replace(content, "http://staticserver.tpl.json", model.Conf.StaticServer, -1)
-	content = strings.Replace(content, "${StaticResourceVersion}", model.Conf.StaticResourceVersion, -1)
-	writePath := strings.TrimSuffix(path, ".tpl")
-	if err = ioutil.WriteFile(writePath, []byte(content), 0644); nil != err {
-		logger.Fatal("replace sw.min.js in [" + path + "] failed: " + err.Error())
+	if util.File.IsExist(path) {
+		data, err := ioutil.ReadFile(path)
+		if nil != err {
+			logger.Fatal("read file [" + path + "] failed: " + err.Error())
+		}
+		content := string(data)
+		content = strings.Replace(content, "http://server.tpl.json", model.Conf.Server, -1)
+		content = strings.Replace(content, "http://staticserver.tpl.json", model.Conf.StaticServer, -1)
+		content = strings.Replace(content, "${StaticResourceVersion}", model.Conf.StaticResourceVersion, -1)
+		writePath := strings.TrimSuffix(path, ".tpl")
+		if err = ioutil.WriteFile(writePath, []byte(content), 0644); nil != err {
+			logger.Fatal("replace sw.min.js in [" + path + "] failed: " + err.Error())
+		}
 	}
 
-	if util.File.IsExist("console/dist/") { // not exist if npm run dev
-		err = filepath.Walk("console/dist/", func(path string, f os.FileInfo, err error) error {
+	if util.File.IsExist("console/dist/") {
+		err := filepath.Walk("console/dist/", func(path string, f os.FileInfo, err error) error {
 			if strings.HasSuffix(path, ".tpl") {
 				data, err := ioutil.ReadFile(path)
 				if nil != err {
